Some keyboard input working

+++ Main/Emulator/bkInput.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-extern uint16_t port0177660;
-extern uint16_t port0177662;
+extern uint8_t port0177660;
+extern uint8_t port0177662;
 extern uint16_t port0177716;
 
 bool OnKey(uint32_t scanCode, bool isKeyUp);
